crud quiz + controle de saisie✅

+++ src/Controller/CompetitionController.php
@@ -3,8 +3,11 @@
 namespace App\Controller;
 
 use App\Entity\Competition;
+use App\Entity\Quiz;
 use App\Form\CompetitionType;
 use App\Repository\CompetitionRepository;
+use App\Repository\QuizRepository;
+use App\Repository\LivreRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -26,7 +29,7 @@ class CompetitionController extends AbstractController
     }
 
     #[Route('/new', name: 'app_competition_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, CompetitionRepository $repo): Response
+    public function new(Request $request, CompetitionRepository $repo, LivreRepository $livreRepo): Response
     {
         $competition = new Competition();
         $form = $this->createForm(CompetitionType::class, $competition);
@@ -34,6 +37,15 @@ class CompetitionController extends AbstractController
 
         if ($form->isSubmitted() && $form->isValid()) {
             $repo->save($competition,true);
+              // Create a new Quiz entity and set its competition and livre properties
+        $quiz = new Quiz();
+        $quiz->setIdCompetition($competition);
+        $livre = $livreRepo->findOneByIdLivre($competition->getIdLivre()); // Replace 1 with the ID of the desired book
+        $quiz->setIdLivre($livre);
+
+        $em = $this->getDoctrine()->getManager();
+        $em->persist($quiz);
+        $em->flush();
             $this->addFlash('success', 'Competition ajoutée avec succés!');
 
             return $this->redirectToRoute('app_competition_index', [], Response::HTTP_SEE_OTHER);
@@ -73,15 +85,20 @@ class CompetitionController extends AbstractController
     }
 
     #[Route('/{idCompetition}', name: 'app_competition_delete', methods: ['POST'])]
-    public function delete(Request $request, Competition $competition, CompetitionRepository $repo): Response
+    public function delete(Request $request, Competition $competition, CompetitionRepository $repo, QuizRepository $quizRepo): Response
     {
         if ($this->isCsrfTokenValid('delete' . $competition->getIdCompetition(), $request->request->get('_token'))) {
             $now = new \DateTime();
         if ($now >= $competition->getDateDebut() && $now <= $competition->getDateFin()) {
-            $this->addFlash('danger', 'Impossible de supprimer la compétition car elle est en cours!');
+            $this->addFlash('danger', 'Impossible de supprimer la compétition car elle est déja ouverte !');
         } else {
-            $repo->remove($competition, true);
-            $this->addFlash('success', 'Compétition supprimée avec succès!');
+            $quiz = $quizRepo->findOneBy(['idCompetition' => $competition->getIdCompetition()]);
+            if ($quiz) {
+                $this->addFlash('danger', 'Impossible de supprimer la compétition vous devez d\'abord supprimer la quiz!');
+            } else {
+                $repo->remove($competition, true);
+                $this->addFlash('success', 'Compétition supprimée avec succès!');
+            }
         }
          
         }
